:sparkles: 서비스 구현

+++ src/main/java/com/ds/ds/domain/group/presentation/data/dto/CreateGroupDto.java
@@ -13,7 +13,7 @@ public class CreateGroupDto {
     private final String groupName;
     private final String groupDescription;
     private final String groupImg;
-    private final Integer groupMaxCount;
+    private final Long groupMaxCount;
     private final Boolean secret;
     private final Optional<String> password;
 }
